refactor: SearchFilter와 SortFilter 분리
